add Trainer, which is a framework for training such classification networks
